requirements.txt
quart
motor
dnspython
bcrypt
Pillow